i changed the way taking of the name of the city/ country and i added the necessary auxiliary fnn and objs of course
App

1. i created and connected the new loader - "Weather
loader"

Weather

1. i added the new hooks "useLoaderData" and "useState"
1.1 the first hook to use data from the new loader.
1.2 the second hook for scrolling the weather forecast
2. changed the way checking for the for if it has
any weather or doesn't
3. added the new fn that returns weather data which
user has choosed.
4. Now the getWeather receive the obj instend of arr and
the url search param to set the name of the city forecast

LocationItem

1. i remove the "useNavigate" hook and use the "Link"
component to navigate

locationLoader

1. changed a little

LocationTable

1. renamed the "getSortedCities" to "getFormatedCities"
and now this fn return the arr of pair "city name,
city weather".
2. getLocationItems receive that arr and now use the name
which is in that arr to pass it as props to it child.

storage

1. created the new fn which returns the weather data
from the localStorage.